Break apart the GUI plugin styles
